refactor(editors/templates): move  wizards and actions related to DAType
